cursor(): clamp negative coords to 0 -- fixes print() wiping the screen
api_cursor took uint8_t, so a negative cursor Y (e.g. cursor(x, sin-based-y))
wrapped to ~255. print()'s auto-scroll then computed final_y = 255 + h - 128
= +133 and ran scrool_screen(133), memmove/memset past the framebuffer -> the
ENTIRE screen cleared to black. Any cart positioning text with a sometimes-
negative y (very common: wavy/bobbing text via sin()) had its frame wiped.

Measured PICO-8 0.2.7a6 (-x): cursor(_,-1)..(_,-8) behave exactly like
cursor(_,0) (no scroll, fb stays full), while cursor(_,120..255) scroll
proportionally. So PC clamps negative cursor coords to 0 on set. Fix: take the
args as int16_t (preserve sign) and clamp negatives to 0 before storing the byte.

Verified on z8headless vs official:
 - cursor(_,-1) print: 16384 (was 0) -- no longer erases
 - scroll sweep -8/-1/0/60/120/128/130/200 all match official exactly
 - Burger Age: 0 px (black) -> ~13700 px, matching official ground truth
 - Lina (prior fix) still 16384, no regression

+++ src/pico8/vm.h
@@ -252,7 +252,7 @@ private:
     void api_serial(int16_t chan, int16_t address, int16_t len);
 
     // Text
-    tup<uint8_t, uint8_t, uint8_t> api_cursor(uint8_t x, uint8_t y, opt<uint8_t> c);
+    tup<uint8_t, uint8_t, uint8_t> api_cursor(int16_t x, int16_t y, opt<uint8_t> c);
     tup<opt<fix32>, opt<fix32> > api_print(opt<rich_string> str, opt<fix32> x, opt<fix32> y, opt<fix32> c);
 
     // Graphics
